Cleaning.

+++ Gamelan/Sources/Utils/benchmark.h
@@ -20,7 +20,7 @@ namespace Benchmark {
 		std::string name;
 		long long start;
 		long long stop;
-		uint32_t threadID;
+		size_t threadID;
 	};
 
 	class Writer {
@@ -99,7 +99,7 @@ namespace Benchmark {
 				std::chrono::time_point<std::chrono::high_resolution_clock> stopPoint = std::chrono::high_resolution_clock::now();
 				long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_startPoint).time_since_epoch().count();
 				long long stop  = std::chrono::time_point_cast<std::chrono::microseconds>(stopPoint).time_since_epoch().count();
-				uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+				size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
 				Writer::get().writeEntry({ m_name, start, stop, threadID });
 			}
 	};
